requirements.txt
flask
anthropic>=0.9.0
python-dotenv
tavily-python
tenacity
httpx==0.27.2
gunicorn==21.2.0
waitress==2.1.2